Fix panic on closures capturing 27+ variables (#809)
Two distinct u8-overflow panics on legal programs:

- allocClosure computed its byte accounting in u8 arithmetic
  (upvalue_count is u8), overflowing at 27 captures (27*8 + 40 > 255)
  and aborting the process on every closure creation. Widen the
  arithmetic to usize.

- addUpvalue @intCast'd the upvalue count into a u8 upvalue_count field,
  panicking past 255 captures. Widen upvalue_count to u16 (matching the
  u16 locals_count and the u16 upvalue index already used in the
  bytecode) so 256+ captures work, and cap the count gracefully at the
  u16 limit with a CompileError instead of a panic.

Serializing upvalue_count as u16 bumps the .sbc format to v5; stale v4
caches are rejected on version mismatch and recompiled.

+++ src/types.zig
@@ -219,7 +219,7 @@ pub const Function = struct {
     constants: std.ArrayList(Value),
     arity: u8,
     locals_count: u16 = 0,
-    upvalue_count: u8 = 0,
+    upvalue_count: u16 = 0,
     is_variadic: bool = false,
     name: ?[]const u8 = null,
     owns_name: bool = false,
